requirements.txt
streamlit
python-dotenv
pypdf
numpy
scikit-learn
sentence-transformers
groq
